Fix bugs in Create-schema component and add not found component

+++ client/src/app/schema-list/schema-list.component.ts
@@ -10,30 +10,20 @@ import {Schema} from "../_models/schema";
 export class SchemaListComponent implements OnInit {
   private schemas: [Schema];
 
-  /* schemas = [{
-     name: 'schema1',
-     fields: [{
-       name: 'boolean'
-     }, {
-       name: 'email'
-     }, {
-       name: 'gender'
-     }]
-   }];
- */
   constructor(private schemaService: SchemaService) {
     schemaService.getSchemasByUserId().subscribe(schemas => this.schemas = schemas);
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   deleteSchema(index) {
-    this.schemaService.remove(this.schemas[index].id)
-      .subscribe(
-        schema => console.log(schema),
-        error => console.log(error)
-      );
-    this.schemas.splice(index, 1);
+    if (confirm("Are you sure to delete this schema?")) {
+      this.schemaService.remove(this.schemas[index].id)
+        .subscribe(
+          schema => console.log(schema),
+          error => console.log(error)
+        );
+      this.schemas.splice(index, 1);
+    }
   }
 }
